Moved onboarding to new file and created splashscreen
